\name{demo}
\alias{demo}
\docType{data}

\title{mrna profile of colorectal cancer, demo}

\description{
The file contains the mRNA profile of CRC samples 
for a demo
}
\usage{
data("demo")
}
\value{
\item{demo}{the demo dataset}
}

\author{Claudio Isella, \email{claudio.isella@ircc.it}}
\keyword{CRISclassifier}

