\name{features}
\alias{features}
\docType{data}
\title{the cris classifier}
\description{
The file contains the genes required for the cris classification
}
\usage{data("features")}

\details{
The \code{features} data will load the gene data required for 
the classification, and it is not intended o be loaded manually 
by the user.
}
\author{Claudio Isella}

\examples{\dontrun{
data("features")
}}
\keyword{CRISclassifier}

